requirements.txt
certifi==2021.5.30
chardet==4.0.0
decorator==4.4.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
moviepy==1.0.3
mss==2.0.22
numpy==1.20.3
opencv-python==4.5.2.54
Pillow==8.2.0
proglog==0.1.9
requests==2.25.1
tqdm==4.61.0
urllib3==1.26.5
wincertstore==0.2
